requirements.txt
torch
torchvision
numpy
pillow
matplotlib
scikit-learn
tqdm
tensorboard
opencv-python
albumentations
